backend/requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
astor==0.8.1
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blis==1.2.0
build==1.2.2.post1
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
cymem==2.0.11
Cython==3.0.12
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
executing==2.2.0
fastapi==0.115.11
fastjsonschema==2.21.1
fhir-core==1.0.0
fhir.resources==8.0.0
fire==0.7.0
fonttools==4.56.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
murmurhash==1.0.12
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
numpy==2.2.3
openai==1.65.5
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.10.0
paddlepaddle==2.6.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
platformdirs==4.3.6
preshed==3.0.9
prompt_toolkit==3.0.50
protobuf==6.30.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
RapidFuzz==3.12.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
scikit-image==0.25.2
scipy==1.15.2
setuptools==76.0.0
shapely==2.0.7
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.1
stringzilla==3.12.3
termcolor==2.5.0
thinc==8.3.4
tifffile==2025.2.18
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
wheel==0.45.1
wrapt==1.17.2
yarg==0.1.9
